Revert "Solve compiler warning due to Vptr.disp change"
This reverts commit d2456cc231dad57e00bd604b72fe92db0fce0ef9.

+++ hphp/runtime/vm/jit/vasm-copy.cpp
@@ -55,7 +55,7 @@ folly::Optional<int32_t> phys_reg_index(Vreg reg) {
  * either copying or lea'ing off another register.  The `base' register is not
  * necessarily a virtual register.
  */
-struct RegInfo { Vreg base; intptr_t disp; };
+struct RegInfo { Vreg base; int32_t disp; };
 
 DEBUG_ONLY std::string show(RegInfo x) {
   return folly::sformat("{} + {}", show(x.base), x.disp);
